Splitting Components By Feature & State
