+++ react-deep-dive/src/components/CoreConcepts.jsx
@@ -1,0 +1,14 @@
+import { CORE_CONCEPTS } from "../data.js";
+import CoreConcept from "../components/CoreConcept.jsx";
+export default function CoreConcepts() {
+  return (
+    <section id="core-concepts">
+      <h2>Core Concepts</h2>
+      <ul>
+        {CORE_CONCEPTS.map((conceptItem) => (
+          <CoreConcept key={conceptItem.title} {...conceptItem} />
+        ))}
+      </ul>
+    </section>
+  );
+}
